Final Importer things
Flip Texture on convert
Check material/textures/meshes are there before trying to make a draw call, not great, should be when caching the groups as it should not be done frame by frame...

Now to do the component Camera with frustum culling

+++ src/modules/ECS/ComponentsRender/ECS_RendererTypes.h
@@ -4,12 +4,12 @@
 #include <src/helpers/MathGeoLib/MathGeoLib.h>
 
 struct MaterialGroup {
-	uint32_t material;
-	std::vector<uint32_t> meshes;
-	std::vector<uint32_t> world_matrices;
+	uint64_t material;
+	std::vector<uint64_t> meshes;
+	std::vector<uint64_t> world_matrices;
 };
 struct RenderGroup {
-	uint32_t gl_state;
+	uint64_t gl_state;
 	std::vector<MaterialGroup> materialgroups;
 };
